3_add_DCR/pyproject.toml
[project]
name = "3-add-dcr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.47.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-iam>=2.20.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
]
